requirements.txt
tensorflow<2.11
Flask
Pillow
gunicorn
sqlalchemy
pymysql